# docs/swagger.yaml
basePath: /
definitions:
  logic.Token:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  model.Vote:
    properties:
      count:
        type: integer
      createdTime:
        type: string
      during:
        type: integer
      id:
        type: integer
      startTime:
        type: integer
      title:
        type: string
      type:
        type: integer
      userId:
        type: integer
      vote_opt:
        items:
          $ref: '#/definitions/model.VoteOpt'
        type: array
    type: object
  model.VoteOpt:
    properties:
      count:
        type: integer
      createdTime:
        type: string
      id:
        type: integer
      name:
        type: string
      percent:
        type: integer
      vote_id:
        type: integer
    type: object
  tools.HttpCode:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:80
info:
  contact: {}
  description: 投票项目的第三个版本呼呼呼
  title: 投票项目
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: 会执行用户登录操作
      parameters:
      - description: 用户名
        in: formData
        name: name
        required: true
        type: string
      - description: 密码
        in: formData
        name: pwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  $ref: '#/definitions/logic.Token'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  $ref: '#/definitions/logic.Token'
              type: object
      summary: 用户登录
      tags:
      - user
  /logout:
    get:
      consumes:
      - application/json
      description: 会执行用户退出操作
      produces:
      - application/json
      responses:
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/tools.HttpCode'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tools.HttpCode'
      summary: 用户退出
      tags:
      - user
  /opt:
    post:
      consumes:
      - application/json
      description: 新增加投票选项
      parameters:
      - description: 投票主ID
        in: body
        name: vote_id
        schema:
          type: integer
      - description: 选项名称
        in: body
        name: name
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tools.HttpCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.HttpCode'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tools.HttpCode'
      summary: 新增投票选项
      tags:
      - opt
    put:
      consumes:
      - application/json
      description: 更新投票的选项
      parameters:
      - description: 投票主题ID
        in: body
        name: id
        schema:
          type: integer
      - description: 投票主题
        in: body
        name: name
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tools.HttpCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.HttpCode'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tools.HttpCode'
      summary: 更新投票选项
      tags:
      - opt
  /opt/:id:
    delete:
      consumes:
      - application/json
      description: 根据ID删除选项
      parameters:
      - description: int valid
        in: path
        minimum: 1
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tools.HttpCode'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/tools.HttpCode'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tools.HttpCode'
      summary: 删除投票选项
      tags:
      - opt
  /opts/{vote_id}:
    get:
      consumes:
      - application/json
      description: 获取投票列表
      parameters:
      - description: int valid
        in: path
        minimum: 1
        name: vote_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.VoteOpt'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.VoteOpt'
                  type: array
              type: object
      summary: 获取投票列表
      tags:
      - opt
  /vote/:id:
    delete:
      consumes:
      - application/json
      description: 根据ID删除投票的主体
      parameters:
      - description: int valid
        in: path
        minimum: 1
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  $ref: '#/definitions/model.Vote'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  $ref: '#/definitions/model.Vote'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  $ref: '#/definitions/model.Vote'
              type: object
      summary: 删除投票主体
      tags:
      - vote
    get:
      consumes:
      - application/json
      description: 根据ID获取某个用户的投票详情
      parameters:
      - description: int valid
        in: path
        minimum: 1
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  $ref: '#/definitions/model.Vote'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  $ref: '#/definitions/model.Vote'
              type: object
      summary: 获取投票详情
      tags:
      - vote
  /vote/basic:
    post:
      consumes:
      - application/json
      description: 新增加投票的主体
      parameters:
      - description: 投票主题
        in: body
        name: title
        schema:
          type: string
      - description: 开始时间 时间戳
        in: body
        name: start_time
        schema:
          type: integer
      - description: 类型 0单选  1多选
        in: body
        name: type
        schema:
          type: integer
      - description: 持续时间 秒
        in: body
        name: during
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tools.HttpCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.HttpCode'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tools.HttpCode'
      summary: 新增投票主体
      tags:
      - vote
    put:
      consumes:
      - application/json
      description: 更新投票的主体
      parameters:
      - description: 投票主题ID
        in: body
        name: id
        schema:
          type: integer
      - description: 投票主题
        in: body
        name: title
        schema:
          type: string
      - description: 开始时间 时间戳
        in: body
        name: start_time
        schema:
          type: integer
      - description: 类型 0单选  1多选
        in: body
        name: type
        schema:
          type: integer
      - description: 持续时间 秒
        in: body
        name: during
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tools.HttpCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tools.HttpCode'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tools.HttpCode'
      summary: 更新投票主体
      tags:
      - vote
  /vote/do:
    post:
      consumes:
      - application/json
      description: 根据用户投票信息执行投票操作
      parameters:
      - description: 投票主体ID
        in: formData
        name: vote_id
        type: integer
      - collectionFormat: csv
        description: 选项ID
        in: formData
        items:
          type: integer
        name: opt[]
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  $ref: '#/definitions/model.Vote'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  $ref: '#/definitions/model.Vote'
              type: object
      summary: 投票
      tags:
      - vote
  /votes:
    get:
      consumes:
      - application/json
      description: 获取投票列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Vote'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/tools.HttpCode'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Vote'
                  type: array
              type: object
      summary: 获取投票列表
      tags:
      - vote
swagger: "2.0"